test: testes regreas de negócio do usuário

+++ nexus_gourmet/backend/services/user_service.py
@@ -1,6 +1,5 @@
-from flask import session
-from backend.models.models import db, User
-from backend.models.enums import Role
+from models.models import db, User
+from models.enums import Role
 from werkzeug.security import generate_password_hash, check_password_hash
  
 class UserService:
@@ -8,6 +7,7 @@ class UserService:
         usuario = self.get_user_by_id(id)
         if not usuario:
             return False, "Usuário não encontrado.", None
+        
         if not check_password_hash(usuario.senha, senha):
             return False, "Senha incorreta.", None
         return True, "Login bem-sucedido.", usuario
@@ -21,13 +21,17 @@ class UserService:
     def listar_usuarios(self):
         return User.query.all()
     
-    def cadastrar_usuario(self, id, nome, senha, cargo):
+    def cadastrar_usuario(self, usuario_logado, id, nome, senha, cargo):
+        if usuario_logado.cargo != Role.ADMINISTRADOR:
+            return False, "Apenas administradores podem cadastrar usuários."
+        
         if self.get_user_by_id(id):
             return False, "ID de usuário já existe."
         try:
             cargo = Role(cargo)
         except ValueError:
             return False, f"Cargo inválido: {cargo}."
+        
         senha_hash = generate_password_hash(senha)
         novo_usuario = User(id=id, nome=nome, senha=senha_hash, cargo=cargo)
         db.session.add(novo_usuario)
@@ -40,10 +44,17 @@ class UserService:
             return False, "Usuário não encontrado."
         return True, usuario
  
-    def editar_usuario(self, id, nome=None, cargo=None):
+    def editar_usuario(self, usuario_logado, id, nome=None, cargo=None):
+        if usuario_logado.cargo != Role.ADMINISTRADOR:
+            return False, "Apenas administradores podem editar usuários."
         usuario = self.get_user_by_id(id)
+
         if not usuario:
             return False, "Usuário não encontrado."
+        
+        if not nome and not cargo:
+            return False, "Nenhum campo fornecido para edição."
+        
         if nome:
             usuario.nome = nome
         if cargo:
@@ -55,8 +66,12 @@ class UserService:
         db.session.commit()
         return True, "Usuário editado com sucesso."    
         
-    def deletar_usuario(self, id):
+    def deletar_usuario(self, usuario_logado, id):
+        if usuario_logado.cargo != Role.ADMINISTRADOR:
+            return False, "Apenas administradores podem excluir usuários."
+        
         usuario = self.get_user_by_id(id)
+
         if not usuario:
             return False, "Usuário não encontrado."
         db.session.delete(usuario)
@@ -67,14 +82,20 @@ class UserService:
         usuario = self.get_user_by_id(id)
         if not usuario:
             return False, "Usuário não encontrado."
+        
         if not check_password_hash(usuario.senha, senha):
             return False, "Senha atual incorreta."
+        
         usuario.senha = generate_password_hash(nova_senha)
         db.session.commit()
         return True, "Senha alterada com sucesso."
     
-    def mudar_cargo(self, id, cargo):
+    def mudar_cargo(self, usuario_logado, id, cargo):
+        if usuario_logado.cargo != Role.ADMINISTRADOR:
+            return False, "Apenas administradores podem alterar cargos."
+        
         usuario = self.get_user_by_id(id)
+
         if not usuario:
             return False, "Usuário não encontrado."
         try:
@@ -85,17 +106,5 @@ class UserService:
         db.session.commit()
         return True, "Cargo alterado com sucesso."
     
-    def transferir_posse(self, id_atual, id_novo):
-        usuario_atual = self.get_user_by_id(id_atual)
-        usuario_novo = self.get_user_by_id(id_novo)
-        if not usuario_atual or not usuario_novo:
-            return False, "Usuário não encontrado."
-        if usuario_atual.cargo != Role.ADMINISTRADOR:
-            return False, "Apenas administradores podem transferir posse."
-        usuario_atual.cargo = Role.USUARIO
-        usuario_novo.cargo = Role.ADMINISTRADOR
-        db.session.commit()
-        return True, "Posse transferida com sucesso."
-    
     def get_user_by_id(self, id):
-        return User.query.get(id)
+        return db.session.get(User, id)
